Created duplicate error checking
